+++ packages/components/src/FastCheckbox.tsx
@@ -74,7 +74,7 @@ const StyledWrapper = styled('div')<{ $color: FastCheckboxColor; $size: number; 
 
   .checkbox-text {
     font-weight: 500;
-    font-size: 0.875rem;
+    font-size: 0.9rem;
     color: ${p => (p.theme.palette.text.primary)};
   }